requirements.txt
requests>=2.31.0
google-generativeai>=0.7.2
edge-tts>=6.1.9
moviepy==1.0.3
Pillow>=10.4.0
numpy==1.26.4
imageio-ffmpeg==0.6.0
setuptools